+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.2.2 2023-08
 
+* [Add plugin event bus `destroy-protyle`](https://github.com/siyuan-note/siyuan/issues/9033)
+* [Add plugin event bus `loaded-protyle-dynamic`](https://github.com/siyuan-note/siyuan/issues/9021)
 
 ## 0.2.1 2023-08-21